+++ src/pages/PanelManagement/PanelManagement.tsx
@@ -105,7 +105,7 @@ export const deleteProposalPanel = (
   setProposalPanels(filtered);
 };
 
-export default function PanelMaintenance() {
+export default function PanelManagement() {
   const { t } = useScopedTranslation();
   const navigate = useNavigate();
   const theme = useTheme();
@@ -237,7 +237,8 @@ export default function PanelMaintenance() {
   return (
     <>
       <PageBannerPMT title={t('page.15.desc')} backBtn={backButton()} fwdBtn={fwdButton()} />
-      <Spacer size={BANNER_PMT_SPACER_MIN} axis={SPACER_VERTICAL} />
+      <Spacer size={BANNER_PMT_SPACER_MIN} axis={{ SPACER_VERTICAL }} />
+      <Spacer size={BANNER_PMT_SPACER_MIN} axis={{ SPACER_VERTICAL }} />
       <Grid
         container
         pr={2}