+++ tests/cypress/e2e/common/common.js
@@ -155,7 +155,7 @@ export const clickCycleConfirm = () => clickButton('cycleConfirmationButton');
 export const clickUserMenu = () => clickButton('usernameMenu');
 export const clickObservationSetup = () => clickButton('addObservationButton');
 export const clickAddObservationEntry = () => clickButton('addObservationButtonEntry');
-export const clickPanelMaintenanceButton = () => clickButton('pmtBackButton');
+export const clickPanelManagementButton = () => clickButton('pmtBackButton');
 export const clickResolveButton = () => clickButton('resolveButton');
 export const clickReviewOverviewButton = () => clickButton('overviewButtonTestId');
 export const clickSave = () => clickButton('saveBtn');
@@ -233,7 +233,7 @@ export const clickUserMenuOverview = () => clickSignINBtns('menuItemOverview', '
 export const clickUserMenuProposals = () => clickSignINBtns('menuItemProposals', '');
 export const clickUserMenuVerification = () => clickSignINBtns('menuItemVerification', '');
 export const clickUserMenuPanels = () =>
-  clickSignINBtns('menuItemPanelSummary', 'PANEL MAINTENANCE');
+  clickSignINBtns('menuItemPanelSummary', 'PANEL MANAGEMENT');
 export const clickUserMenuReviews = () => clickSignINBtns('menuItemReviews', 'REVIEW PROPOSALS');
 export const clickUserMenuDecisions = () =>
   clickSignINBtns('menuItemReviewDecisions', 'REVIEW DECISIONS');