+++ CHANGELOG.md
@@ -9,7 +9,7 @@ UNRELEASED
 * Changed: Display of Add target tab on target page
 * Added: Pulsar timing beam functionality when creating/editing a target
 * Adds PMT to User Guide structure.
-* Added: Calibration page
+* Changed Panel Maintenance to Panel Management.
 
 7.0.0
 ----------